+++ odoo/local-src/bso_delivery/models/__init__.py
@@ -1,9 +1,9 @@
-from . import sale_dealsheet
 from . import delivery_checklist
-from . import delivery_project
-from . import sale_order
-from . import jira_product_template
-from . import delivery_project_line
+from . import delivery_doc
 from . import service_point
 from . import stock_picking
-from . import jira_project
+
+from . import base_delivery
+from . import sale_project
+from . import purchase_project
+from . import jira